#8 Small refactor on base forms properties, methods... Preparing the code for a new winform for Settings
