+++ Library/CommonScripts/View/Base/BaseInnerForm.resx
